processor/requirements.txt
pandas
boto3
python-dotenv
scikit-learn
